fix(k8s-install): fix all shellcheck warning

+++ test/cert_export__invalid_target_cert_file.sh
@@ -12,15 +12,15 @@ function test {
   # Create a certificate
   cert test.domain.com > /dev/null
   # Attempt to export the key on a directory
-  mkdir -p $KEYSER_VAULT_DIR/some/target/com.domain.test.cert.pem
+  mkdir -p "$KEYSER_VAULT_DIR"/some/target/com.domain.test.cert.pem
   res=$(cert_export test.domain.com "$KEYSER_VAULT_DIR"/some/target)
-  [[ $? == 0 ]] && return 1
+  [[ $? == 1 ]] || return 1
   echo "$res" | grep 'Target certificate is not a file.' > /dev/null || return 1
   # Attempt to export the key on a file
   rm -r "$KEYSER_VAULT_DIR"/some/target/com.domain.test.cert.pem
   touch "$KEYSER_VAULT_DIR"/some/target/com.domain.test.cert.pem
   res=$(cert_export test.domain.com "$KEYSER_VAULT_DIR"/some/target)
-  [[ $? == 0 ]] && return 1
+  [[ $? == 1 ]] || return 1
   echo "$res" | grep 'Target certificate file already exists.' > /dev/null || return 1
 }
