+++ test/cert_export__invalid_target_key_file.sh
@@ -14,13 +14,13 @@ function test {
   # Attempt to export the key on a directory
   mkdir -p "$KEYSER_VAULT_DIR"/some/target/com.domain.test.key.pem
   res=$(cert_export test.domain.com "$KEYSER_VAULT_DIR"/some/target)
-  [[ $? == 0 ]] && return 1
+  [[ $? == 1 ]] || return 1
   echo "$res" | grep 'Target key is not a file.' > /dev/null || return 1
   # Attempt to export the key on a file
   rm -r "$KEYSER_VAULT_DIR"/some/target/com.domain.test.key.pem
   touch "$KEYSER_VAULT_DIR"/some/target/com.domain.test.key.pem
   res=$(cert_export test.domain.com "$KEYSER_VAULT_DIR"/some/target)
-  [[ $? == 0 ]] && return 1
+  [[ $? == 1 ]] || return 1
   echo "$res" | grep 'Target key file already exists.' > /dev/null || return 1
 }
